refector: API 폴더 구조 재정비 및 서비스 레이어 분리
