+++ src/app/(protected)/square/page.tsx
@@ -1,8 +1,8 @@
-// API 연동
-import { getSharedDreamListServer } from "@/lib/api/queries";
-
 // components
 import SharedDreamList from "./_components/SharedDreamList";
+
+// services
+import { getSharedDreamListServer } from "@/services/server";
 
 export default async function SharedDreamFeedPage() {
   const res = await getSharedDreamListServer();
